requirements.txt
Django~=4.2.9
djangorestframework~=3.14.0
apscheduler~=3.10.4
djongo~=1.3.6
sqlparse==0.2.4
django-elasticsearch-dsl